Update debug code to not use Startup /2
