--- aspnetcore/test/troubleshoot/code/5.x/Program.cs
@@ -1,75 +1,0 @@
-// This is a copy/paste from the previous version embedded in the MD file, used only for productive
-// GitHub diff
-public void Configure(IApplicationBuilder app, IHostingEnvironment env,
-    IConfiguration config)
-{
-    if (env.IsDevelopment())
-    {
-        app.Run(async (context) =>
-        {
-            var sb = new StringBuilder();
-            var nl = System.Environment.NewLine;
-            var rule = string.Concat(nl, new string('-', 40), nl);
-            var authSchemeProvider = app.ApplicationServices
-                .GetRequiredService<IAuthenticationSchemeProvider>();
-
-            sb.Append($"Request{rule}");
-            sb.Append($"{DateTimeOffset.Now}{nl}");
-            sb.Append($"{context.Request.Method} {context.Request.Path}{nl}");
-            sb.Append($"Scheme: {context.Request.Scheme}{nl}");
-            sb.Append($"Host: {context.Request.Headers["Host"]}{nl}");
-            sb.Append($"PathBase: {context.Request.PathBase.Value}{nl}");
-            sb.Append($"Path: {context.Request.Path.Value}{nl}");
-            sb.Append($"Query: {context.Request.QueryString.Value}{nl}{nl}");
-
-            sb.Append($"Connection{rule}");
-            sb.Append($"RemoteIp: {context.Connection.RemoteIpAddress}{nl}");
-            sb.Append($"RemotePort: {context.Connection.RemotePort}{nl}");
-            sb.Append($"LocalIp: {context.Connection.LocalIpAddress}{nl}");
-            sb.Append($"LocalPort: {context.Connection.LocalPort}{nl}");
-            sb.Append($"ClientCert: {context.Connection.ClientCertificate}{nl}{nl}");
-
-            sb.Append($"Identity{rule}");
-            sb.Append($"User: {context.User.Identity.Name}{nl}");
-            var scheme = await authSchemeProvider
-                .GetSchemeAsync(IISDefaults.AuthenticationScheme);
-            sb.Append($"DisplayName: {scheme?.DisplayName}{nl}{nl}");
-
-            sb.Append($"Headers{rule}");
-            foreach (var header in context.Request.Headers)
-            {
-                sb.Append($"{header.Key}: {header.Value}{nl}");
-            }
-            sb.Append(nl);
-
-            sb.Append($"WebSockets{rule}");
-            if (context.Features.Get<IHttpUpgradeFeature>() != null)
-            {
-                sb.Append($"Status: Enabled{nl}{nl}");
-            }
-            else
-            {
-                sb.Append($"Status: Disabled{nl}{nl}");
-            }
-
-            sb.Append($"Configuration{rule}");
-            foreach (var pair in config.AsEnumerable())
-            {
-                sb.Append($"{pair.Path}: {pair.Value}{nl}");
-            }
-            sb.Append(nl);
-
-            sb.Append($"Environment Variables{rule}");
-            var vars = System.Environment.GetEnvironmentVariables();
-            foreach (var key in vars.Keys.Cast<string>().OrderBy(key => key,
-                StringComparer.OrdinalIgnoreCase))
-            {
-                var value = vars[key];
-                sb.Append($"{key}: {value}{nl}");
-            }
-
-            context.Response.ContentType = "text/plain";
-            await context.Response.WriteAsync(sb.ToString());
-        });
-    }
-}
